% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/online_emulate_maize.R
\name{online_emulate_maize}
\alias{online_emulate_maize}
\title{Perform online prediction for Maize outputs from APSIMX}
\usage{
online_emulate_maize(train, test, pred_var, pred_type, method, local_dist = 30)
}
\description{
This function performs online prediction on next generation
APSIM data. The 'train' data should be in the same format as
the 'test' data and include the standard weather varibles. 
Currently, this function works best for Maize model outputs,
but may be applicable to other outputs. Variables with
non-zero outputs during days of the year when the crop is 
not growing will be predicted to be 0 and likely innaccurate.
}
\details{
\code{train} A data.frame of training data

\code{test} A data.frame of data to perform online prediction

\code{pred_var} Any Maize model APSIM output variable

\code{pred_type} One of "no_update", "online_total_local", 
"online_total_full", "online_change_local", "online_change_full"

\code{method} One of "lm", "ar", "gam", "rf", "nnet"

\code{local_dist} Number of sliding window observations for local training
}
\examples{
data("training")
data("testing")
on <- online_emulate_maize(training, testing, pred_var = "Maize.AboveGround.Wt", pred_type = "online_total_local", method = "gam")
plot(testing$Date, on, type = "l")
}
