% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\docType{data}
\name{simulation}
\alias{simulation}
\title{Simulation}
\format{A data.frame with the following columns:
  
  \code{CheckpointID}: Unique integer id of observation}
\usage{
simulation
}
\description{
35 Year APSIM Next Generation Simulation to use for Online Prediction
}
\keyword{datasets}
